Refactor : 예외 처리 패키지 구조 변경
예외 처리 패키지 구조를 변경하였습니다.
