+++ src/main/java/nbbang/com/nbbang/global/error/GlobalErrorResponseMessage.java
@@ -1,15 +1,16 @@
-package nbbang.com.nbbang.global.response;
+package nbbang.com.nbbang.global.error;
 
-public class GlobalResponseMessage {
+public class GlobalErrorResponseMessage {
 
     public static final String ILLEGAL_ARGUMENT_ERROR = "올바르지 않은 요청 데이터입니다.";
     public static final String INTERNAL_SERVER_ERROR = "서버 내부 에러";
     public static final String DB_ERROR = "데이터베이스 에러";
 
     public static final String ILLEGAL_TYPE_CONVERSION_ERROR = "값 타입 변환에 실패하였습니다. 올바른 값 타입인지 확인해 주세요.";
+
     public static final String NOT_OWNER_ERROR = "방장이 아닙니다.";
     public static final String NOT_PARTY_MEMBER_ERROR = "파티의 멤버가 아닙니다.";
-    public static final String UNAUTHORIZED_ERROR = "로그인이 필요합니다.";
 
+    public static final String UNAUTHORIZED_ERROR = "로그인이 필요합니다.";
 
 }
