+++ src/main/java/nbbang/com/nbbang/global/error/exception/IllegalPlaceException.java
@@ -1,4 +1,4 @@
-package nbbang.com.nbbang.global.exception;
+package nbbang.com.nbbang.global.error.exception;
 
 /**
  * Place 값을 NONE, SINCHON, YEONHUI, CHANGCHEON 이외의 것으로 입력했을 경우에 발생
