+++ src/main/java/nbbang/com/nbbang/global/error/exception/UnauthorizedException.java
@@ -1,4 +1,4 @@
-package nbbang.com.nbbang.global.exception;
+package nbbang.com.nbbang.global.error.exception;
 
 public class UnauthorizedException extends RuntimeException {
 }
